+++ scripts/effect-engine/effect-engine.js
@@ -1,6 +1,7 @@
 import { componentRegistry } from "./component-registry.js";
 import { conditionComponent } from "./components/condition-component.js";
 import { modifierComponent } from "./components/modifier-component.js";
+import { persistentDamageComponent } from "./components/persistent-damage-component.js";
 
 let initialized = false;
 
@@ -9,6 +10,7 @@ export function initializeEffectEngine() {
 
   componentRegistry.register(conditionComponent);
   componentRegistry.register(modifierComponent);
+  componentRegistry.register(persistentDamageComponent);
 
   initialized = true;
 }